backends: implement Linux hidraw backend
This patch adapts the old code from the legacy branch and implements a
backend for our abstractions. It makes a very big improvement, we now
use a proper tree data type, provided by treelib. This tree data type
has some slightly weird corners in its API but works pretty well
overall.
The code is not too complicated, I think. We look for devices in udev,
and watch changes to update the tree.

Signed-off-by: Filipe Laíns <lains@riseup.net>