backend: add a lock to access device read/write operation
We want to be able to poll devices for events in a parallel thread as
the thread dispatching user actions. Both threads will have to access
the device simultaneously, so we need a lock. We can't simply add a lock
on read/write operations because sometimes, actually, most times, we
will send a message to the device and wait for a response, which needs
to happen on the same lock. We could add a "commad" action that does a
write and a consecutive read, but there are some cases where we would
actually want to wait for a specific response -- the read could read for
example a mouse packet, instead of the protocol packet we want. So, we
need more control, the solution is to change IODevice so that it
implements a context manager which will lock and return an interface,
and once we are done with it, the lock will be released.

Example Usage:

    io_device = DummyDevice()
    with io_device as d:
        d.write(...)

        # wait for the correct response (first byte is 0x99)
        while (response := d.read())[0] != 0x99:
            pass

        ...

Signed-off-by: Filipe Laíns <lains@riseup.net>